Move `entries/` directory higher in source tree
