+++ src/core/index.ts
@@ -1,5 +1,4 @@
 export * from './asset';
-export * from './entries';
 export * from './entry';
 export * from './reference-error';
 export * from './request';
