+++ src/entries/scoped.entry.ts
@@ -1,7 +1,7 @@
-import { CxEntry } from '../entry';
-import { CxRequest } from '../request';
-import { CxRequestMethod } from '../request-method';
-import { CxValues } from '../values';
+import { CxEntry } from '../core/entry';
+import { CxRequest } from '../core/request';
+import { CxRequestMethod } from '../core/request-method';
+import { CxValues } from '../core/values';
 
 /**
  * Creates scoped context value definer.
